Add a generic nodejs installer from the tarballs they release - for #25
The whole nodesource insanity + the non-lts versions being available everywhere that's not debian, means that this is kind of the most
sensible cross-platform installation of nodejs. Also means that you can switch versions by switching out the symlinks.

Should check the shashums in the future, but this took long enough with my lack of ansible knowledge. I hate the jinja filter stuff, don't
know if there's a better way to do the {{ var | map | list | first }} stuff, but that list is damn verbose.