Ansible for installing postgres and docker + indents
I understand why people just put over shell scripts and do simple ansible to check on this.
This was horrible. Really not clear at all what incantations of when + jinja2 that would work.
Also didn't help that postgres didn't have a clear way to figure out if this was needed, so needed something sequential.
Could have easily done a creates: in the initdb step, but this wouldn't have worked for the createuser:
they had to happen at the same time, hence the awkwardness.