Ansible branch now feature complete - #25
Although far far from perfect. But it's a good stepping stone to improve on.
Things are more modular, can be re-run on demand, and should at least work pretty well for Arch.
Whether I can be assed to do the same for Debian (an OS I increasingly am starting to dislike),
and Gentoo (which requires solving other problems first), will be left as separate issues.