node setup that gives me latest verified lts + node2
moved craziness out into a more sensible requests script and now the
ansible is pretty sane. would have been fine with using packages for
nodejs, but it doesn't bootstrap npm, and then when i downgrade npm it
starts confusing the other deps that npm pulled in (http-parser which is
a node lib).

Think it's probably best that npm and os package managers never mix.